Settings: Fix usage of ChooseLockSettingsHelper
The fragment was not being passed, so the ChooseLock* fragments are not
informed if authentication fails (i.e. the cancel button is pushed) and
continue on as if it was successful. This wasn't noticed because this
code isn't used in the normal path where authentication happens before
entering the preferences menu to choose the authentication method.

The authentication support built-in to these fragments may even be dead
code, but it should work correctly if it's there because it may be the
basis of new code.

Change-Id: I859b4e78a55b8a1d3709826824292923be1fd3d9